chore: rename governance components
