+++ packages/shared/routes/dashboard/governance/views/index.js
@@ -1,8 +1,9 @@
 export { default as GovernanceDashboard } from './GovernanceDashboard.svelte'
-export { default as GovernanceEventDetails } from './GovernanceEventDetails.svelte'
-export { default as GovernanceEventInfo } from './GovernanceEventInfo.svelte'
-export { default as GovernanceEventResults } from './GovernanceEventResults.svelte'
-export { default as GovernanceEvents } from './GovernanceEvents.svelte'
-export { default as GovernanceEventStats } from './GovernanceEventStats.svelte'
-export { default as GovernanceInfo } from './GovernanceInfo.svelte'
 export { default as GovernanceVotingWeight } from './GovernanceVotingWeight.svelte'
+export { default as GovernanceInfo } from './GovernanceInfo.svelte'
+export { default as GovernanceEvents } from './GovernanceEvents.svelte'
+
+export { default as GovernanceEvent } from './GovernanceEvent.svelte'
+export { default as GovernanceEventInfo } from './GovernanceEventInfo.svelte'
+export { default as GovernanceEventStats } from './GovernanceEventStats.svelte'
+export { default as GovernanceEventResults } from './GovernanceEventResults.svelte'
